aiven/requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
importlib-metadata==4.0.1
iniconfig==1.1.1
kafka-python==2.0.2
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
SQLAlchemy==1.2.7
toml==0.10.2
typing-extensions==3.10.0.0
zipp==3.4.1
